tests/wpt/web-platform-tests/webxr/xrDevice_requestSession_immersive.https.html

<!DOCTYPE html>
<body>
  <script src=/resources/testharness.js></script>
  <script src=/resources/testharnessreport.js></script>
  <script src="resources/webxr_util.js"></script>
  <canvas></canvas>
  <script>
    xr_session_promise_test(
      "Tests requestSession resolves when supported",
      (session) => {
        assert_not_equals(session, null);
      }, { supportsImmersive:true }, { mode: 'immersive-vr' });
  </script>
</body>
